Extracted method and linearization of variables declaration
